replace both add_exam_to_group and delete_exam_from_;2~group to group_exam

+++ egegrouper/model_sqlite3.py
@@ -235,43 +235,35 @@ class GrouperModelSqlite3(GrouperModel):
         WHERE group_id = ? """, [group_id, ])
         self.conn.commit()
 
-    def add_exam_to_group(self, exam_id, group_id):
-        """Add examination into group.
+    def group_exam(self, exam_id, group_ids, placed_in):
+        """Add and delete examination to and from groups.
 
         Parameters
         ----------
         exam_id : str
-            Examination ID.
-        group_id : str
-            Group ID.
+            Examination identifier.
+        group_ids : list of str
+            Group identifiers.
+        placed_in : list of bool
+            True for examinations to be placed in groups. Length of group_ids must be equal to length of placed_in.
 
         """
-        try:
-            self.c.execute("""
-            INSERT OR REPLACE INTO group_element
-            VALUES (?, ?) """, [exam_id, group_id, ])
-            self.conn.commit()
-        except sqlite3.IntegrityError:
-            print('Error: no such examination or group')
-
-    def delete_exam_from_group(self, exam_id, group_id):
-        """Delete examination from group.
-
-        Parameters
-        ----------
-        exam_id : str
-            Examination ID.
-        group_id : str
-            Group ID.
-
-        """
-        try:        
-            self.c.execute("""
-            DELETE FROM group_element
-            WHERE exam_id = ? AND group_id = ? """, [exam_id, group_id, ])
-            self.conn.commit()
-        except sqlite3.IntegrityError:
-            print('Error: no such examination or group')
+        for (group_id, p) in zip(group_ids, placed_in):
+            if p:
+                try:
+                    self.c.execute("""
+                    INSERT OR REPLACE INTO group_element
+                    VALUES (?, ?) """, [exam_id, group_id, ])
+                except sqlite3.IntegrityError:
+                    pass
+            else:
+                try:
+                    self.c.execute("""
+                    DELETE FROM group_element
+                    WHERE exam_id = ? AND group_id = ? """, [exam_id, group_id, ])
+                except sqlite3.IntegrityError:
+                    pass
+        self.conn.commit()
 
     def where_exam(self, exam_id):
         """Return description of groups where examination in or not in.
